# frontend/rodia-monorepo/apps/mobile/.orval/openapi.raw.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/shipper/inquiries": {
      "post": {
        "tags": [
          "Inquiry"
        ],
        "summary": "화주 문의 생성",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquiryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "생성된 문의 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/push-tokens/me": {
      "post": {
        "tags": [
          "FCM Token"
        ],
        "summary": "FCM 토큰 등록/갱신",
        "operationId": "upsertMyToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FcmTokenUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "토큰 등록 완료"
          }
        }
      },
      "delete": {
        "tags": [
          "FCM Token"
        ],
        "summary": "FCM 토큰 비활성화",
        "operationId": "deactivateMyToken",
        "parameters": [
          {
            "name": "fcmToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "토큰 비활성화 완료"
          }
        }
      }
    },
...
    "/api/driver/inquiries": {
      "post": {
        "tags": [
          "Inquiry"
        ],
        "summary": "기사 문의 생성",
        "operationId": "createByDriver",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquiryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "생성된 문의 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/trucks/{truckId}/approval": {
      "patch": {
        "tags": [
          "admin-truck-controller"
        ],
        "operationId": "updateApproval",
        "parameters": [
          {
            "name": "truckId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruckApprovalUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/inquiries/{inquiryId}/answer": {
      "patch": {
        "tags": [
          "Inquiry"
        ],
        "summary": "관리자 문의 답변",
        "operationId": "answer",
        "parameters": [
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquiryAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "답변된 문의 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reference/truck-specs": {
      "get": {
        "tags": [
          "reference-catalog-controller"
        ],
        "operationId": "getTruckSpecs",
        "parameters": [
          {
            "name": "vehicleType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TruckSpecReferenceResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reference/pricing-rates": {
      "get": {
        "tags": [
          "reference-catalog-controller"
        ],
        "operationId": "getPricingRates",
        "parameters": [
          {
            "name": "vehicleType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PricingRateReferenceResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reference/pricing-rate": {
      "get": {
        "tags": [
          "reference-catalog-controller"
        ],
        "operationId": "getPricingRate",
        "parameters": [
          {
            "name": "distanceKm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "vehicleType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRateReferenceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TruckResponse": {
        "type": "object",
        "properties": {
          "truckId": {
            "type": "integer",
            "format": "int64"
          },
          "driverId": {
            "type": "integer",
            "format": "int64"
          },
          "vehicleType": {
            "type": "string"
          },
          "vehicleBodyType": {
            "type": "string"
          },
          "tonnage": {
            "type": "number"
          },
          "maxWeight": {
            "type": "number"
          },
          "maxVolume": {
            "type": "number"
          },
          "cargoLength": {
            "type": "number"
          },
          "cargoWidth": {
            "type": "number"
          },
          "cargoHeight": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          },
          "approvalStatus": {
            "type": "string"
          },
          "reviewMemo": {
            "type": "string"
          },
          "selected": {
            "type": "boolean"
          },
          "insurance": {
            "type": "string"
          },
          "odometerKm": {
            "type": "number"
          },
          "lastInspectionDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InquiryCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "content": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "content",
          "title"
        ]
      },
      "InquiryResponse": {
        "type": "object",
        "properties": {
          "inquiryId": {
            "type": "integer",
            "format": "int64"
          },
          "shipperId": {
            "type": "integer",
            "format": "int64"
          },
          "driverId": {
            "type": "integer",
            "format": "int64"
          },
          "writerType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "answeredByAdminId": {
            "type": "integer",
            "format": "int64"
          },
          "answeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FcmTokenUpsertRequest": {
        "type": "object",
        "properties": {
          "deviceType": {
            "type": "string",
            "enum": [
              "ANDROID",
              "IOS",
              "WEB"
            ]
          },
          "fcmToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "deviceType",
          "fcmToken"
        ]
      },
...
      "TruckApprovalUpdateRequest": {
        "type": "object",
        "properties": {
          "approved": {
            "type": "boolean"
          }
        }
      },
...
      "InquiryAnswerRequest": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "answer"
        ]
      },
...
      "TruckSpecReferenceResponse": {
        "type": "object",
        "properties": {
          "vehicleType": {
            "type": "string"
          },
          "vehicleTypeKr": {
            "type": "string"
          },
          "vehicleBodyType": {
            "type": "string"
          },
          "categoryKr": {
            "type": "string"
          },
          "tonnage": {
            "type": "number"
          },
          "maxWeight": {
            "type": "number"
          },
          "maxWeightDisplay": {
            "type": "string"
          },
          "maxVolume": {
            "type": "number"
          },
          "cargoLengthCm": {
            "type": "integer",
            "format": "int32"
          },
          "cargoWidthCm": {
            "type": "integer",
            "format": "int32"
          },
          "cargoHeightCm": {
            "type": "integer",
            "format": "int32"
          },
          "palletCount": {
            "type": "integer",
            "format": "int32"
          },
          "palletStandardMm": {
            "type": "string"
          },
          "doorPosition": {
            "type": "string"
          },
          "sourceName": {
            "type": "string"
          }
        }
      },
      "PricingRateReferenceResponse": {
        "type": "object",
        "properties": {
          "rangeKey": {
            "type": "string"
          },
          "minDistanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "maxDistanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "vehicleType": {
            "type": "string"
          },
          "baseRateWon": {
            "type": "integer",
            "format": "int32"
          },
          "sourceName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}